pyproject.toml
[project]
name = "intrinsic-reward-agents"
version = "0.0.1"
description = "Persona/emotion vectors as intrinsic reward for RL-finetuning LLM agents in multi-agent environments."
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Zizhao Hu", email = "zizhaohu3@gmail.com" }]
license = { text = "MIT" }

dependencies = [
    "torch>=2.3",
    "transformers>=4.44",
    "accelerate>=0.33",
    "peft>=0.12",
    "trl>=0.11",
    "datasets>=2.20",
    "numpy>=1.26",
    "pyyaml>=6.0",
    "wandb>=0.17",
    "tqdm>=4.66",
    "rich>=13.7",
]

[project.optional-dependencies]
concordia = ["gdm-concordia>=1.8"]
dev = ["pytest>=8.0", "ruff>=0.6", "pre-commit>=3.8"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/intrinsic_agents"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
